+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Versioning].
   it easier to configure.
 - (`207ad89`) _(Breaking)_ Rework the `no-top-level-variables` rule to focus on
   variables being mutable instead of other side effects.
+- (`a3c513e`) Allow for configuring which classes can be instantiated at the top
+  level for `no-top-level-side-effects`.
 
 ## [Unreleased v2]